refactor: merge table selection plugin files
